% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEplotPlus.R
\name{calc_ALE_varimp}
\alias{calc_ALE_varimp}
\title{calc_ALE_varimp}
\usage{
calc_ALE_varimp(X, MODEL, K = 40, pred.fun)
}
\arguments{
\item{X}{the data used to train the model, without predicted var}

\item{MODEL}{the model}

\item{K}{the number of bins split to evaluate the ALE Plot see ALEPLOT package}

\item{pred.fun}{the prediction function required by ALEPLOT package see details}
}
\value{
a vector of estimated effect intensities for each point
}
\description{
given the X coordinates of the two predictors used to generate an ALEPlot
and the ALEPlot info, calculate the est effect for each point for the
input dataset
}
