% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEplotPlus.R
\name{calc_ALE_varimp_df}
\alias{calc_ALE_varimp_df}
\title{calc_ALE_varimp}
\usage{
calc_ALE_varimp_df(X, X.MODEL, K = 40, pred.fun)
}
\arguments{
\item{X}{the data used to train the model, without predicted var}

\item{X.MODEL}{the model}

\item{K}{the number of bins split to evaluate the ALE Plot see ALEPLOT package}

\item{pred.fun}{the prediction function required by ALEPLOT package see details}
}
\value{
a matrix of estimated effects for each point, for each predictor
}
\description{
An internal helper function designed to assist in calculating variable importance scores
using the ALEPlot method.
The method works as such. Use the ALE method to estimate the effect on prediction
for each datapoint. Other functions summarize this data matrix in various ways
to produce variable importance scores of different flavors
}
