% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEplotPlus.R
\name{calc_ALE_varimps_AUROC}
\alias{calc_ALE_varimps_AUROC}
\title{calc_ALE_varimps_AUROC}
\usage{
calc_ALE_varimps_AUROC(X, X.MODEL, pred.fun, K = 40, predictions, truevals)
}
\arguments{
\item{X}{the data used to train the model, without predicted var}

\item{X.MODEL}{the model}

\item{pred.fun}{the prediction function required by ALEPLOT package see details}

\item{K}{the number of bins split to evaluate the ALE Plot see ALEPLOT package}

\item{predictions}{vector of model predictions}

\item{truevals}{a vector of true model predictions}
}
\description{
Generates variable importance scores for 1D ALEPlots. This variant is designed
to estimate the effect on accuracy of the prediction by adjusting the model predictions
by the estimated effects, then comparing the differences in AUROC to understand
how much worse the model would be if the variable were factored out.
}
