% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEplotPlus.R
\name{calc_ALE_varimps_mean}
\alias{calc_ALE_varimps_mean}
\title{calc_ALE_varimps_mean}
\usage{
calc_ALE_varimps_mean(X, X.MODEL, pred.fun, K = 40)
}
\arguments{
\item{X}{the data used to train the model, without predicted var}

\item{X.MODEL}{the model}

\item{pred.fun}{the prediction function required by ALEPLOT package see details}

\item{K}{the number of bins split to evaluate the ALE Plot see ALEPLOT package}
}
\description{
Generates variable importance scores for 1D ALEPlots. takes the absolute value of the mean
}
