% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEplotPlus.R
\name{calc_vals}
\alias{calc_vals}
\title{calc_vals}
\usage{
calc_vals(x_points, y_points, x_vec, y_vec, z_vec)
}
\arguments{
\item{x_points}{the positions of predictor 1}

\item{y_points}{the positions of predictors2}

\item{x_vec}{the vector of x_positions produced by ALEPlot}

\item{y_vec}{the matrix of z positions produced by ALEPlot}
}
\value{
a vector of estimated interaction intensities for each point
}
\description{
given the X and Y coordinates of the two predictors used to generate an ALEPlot
and the ALEPlot info, calculate the interaction effect for each point for the
input dataset
}
