% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEplotPlus.R
\name{calc_vals_1D}
\alias{calc_vals_1D}
\title{calc_vals_1D}
\usage{
calc_vals_1D(x_points, x_vec, z_vec)
}
\arguments{
\item{x_points}{the positions of predictor 1}

\item{x_vec}{the vector of x_positions produced by ALEPlot}

\item{z_vec}{the matrix of z positions produced by ALEPlot}
}
\value{
a vector of estimated effect intensities for each point
}
\description{
given the X coordinates of the two predictors used to generate an ALEPlot
and the ALEPlot info, calculate the est effect for each point for the
input dataset
}
