% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEplotPlus.R
\name{color.bar}
\alias{color.bar}
\title{color.bar}
\usage{
color.bar(
  lut,
  min,
  max = -min,
  nticks = 4,
  ticks = seq(min, max, len = nticks),
  title = "",
  label = "Effect on Prediction"
)
}
\arguments{
\item{lut}{A color palette (a vector of color values) used to create the color bar.}

\item{min}{The minimum value of the scale, placed at the bottom of the color bar.}

\item{max}{The maximum value of the scale, placed at the top of the color bar (default is -min).}

\item{nticks}{The number of tick marks on the color bar (default is 4).}

\item{ticks}{A numeric vector of positions for tick marks on the color bar (default is `seq(min, max, len = nticks)`).}

\item{title}{A title for the color bar (default is an empty string).}
}
\value{
Invisible NULL. The function is called for its side effect: producing a plot.
}
\description{
Plot a Color Bar
}
\details{
This function plots a color bar with a given color scale and axis labels.
}
\examples{
color.bar(colorRampPalette(c("light green", "yellow", "orange", "red"))(100), -1)

}
\keyword{internal}
