% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEplotPlus.R
\name{create_2D_ALEs}
\alias{create_2D_ALEs}
\title{create_2D_ALEs}
\usage{
create_2D_ALEs(X, X.MODEL, pred.fun, K = 40, savedir = "ALE2DPlots")
}
\arguments{
\item{X}{a dataframe with the same structure as that used to train the model}

\item{X.MODEL}{a model you are interrogating}

\item{pred.fun}{the prediction function used to generate ALEPlot...see doc there}

\item{K}{how many subdivisions to estimate the int coef. for, see ALEPlot}

\item{savedir}{where to save all the ALEPlots generated}
}
\value{
saves all combinations of plots
}
\description{
Produces and saves ALEPlots but with better colors and plotted values which
overall makes it significantly more legible
}
