% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEplotPlus.R
\name{facet.themes}
\alias{facet.themes}
\title{a theeme for faceted plots which does not include a legend for nice tiling.}
\usage{
facet.themes()
}
\description{
a theeme for faceted plots which does not include a legend for nice tiling.
}
