% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEplotPlus.R
\name{find_ALE_interacts}
\alias{find_ALE_interacts}
\title{find_ALE_interacts}
\usage{
find_ALE_interacts(X, X.MODEL, pred.fun, K = 40)
}
\arguments{
\item{X}{the dataframe of predictors used to train model}

\item{X.MODEL}{the machine learning model use}

\item{pred.fun}{the prediction function used with ALEPlot, see ALEPlot for details}

\item{K}{the number of bins in which to assess interactions, see ALEPlot for detaal}
}
\value{
a datagrame estimated interaction strengths. Estimated by multiplying each
       point by its estimated interaction effect
}
\description{
Return a dataframe of the estimate interactions between predictors in an MLModel
using their ALEPlots
}
