% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEplotPlus.R
\name{format_sigfig}
\alias{format_sigfig}
\title{Format a number to three significant figures}
\usage{
format_sigfig(x)
}
\arguments{
\item{x}{A numeric value to be formatted.}
}
\value{
A character string representing the formatted number.
}
\description{
This function formats a number to three significant figures.
It handles both small and large numbers appropriately.
}
\keyword{internal}
