% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEplotPlus.R
\name{interacts2_histogram}
\alias{interacts2_histogram}
\title{interacts2_histogram}
\usage{
interacts2_histogram(ALE_interacts, bins = 1000, ncol = 1)
}
\arguments{
\item{bins}{the number of histogram bins}

\item{ncol}{how many columns in the facetplot}

\item{the}{output of findALEinteractions}
}
\description{
summarizes the information from the findALEinteractions
function into a set of histograms for comparison, just a different
way to do the data.
}
