% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEplotPlus.R
\name{mk.colors_ale}
\alias{mk.colors_ale}
\title{mk.color_ale}
\usage{
mk.colors_ale(nclrs = 30, clr_st = "Spectral", even = T, mtrx, logd = F)
}
\arguments{
\item{nclrs}{how many colors}

\item{clr_st}{which color set, divergent preferred}

\item{even}{whether to set  the middle color to 0}

\item{mtrx}{whether to set  the middle color to 0}

\item{logd}{whether color scale should be logged}
}
\value{
a list of the color codes, and their accompanying breaks
}
\description{
makes_colors for the aleplots
}
