% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEplotPlus.R
\name{mkALEplots}
\alias{mkALEplots}
\title{ALEPlots are a method to visualize the relationship between individual predictors and an outcome within ordinarily
difficult to interpret non-linear machine learning models. This function produces a list of plots with an aleplot
for each predictor within an ML model.}
\usage{
mkALEplots(X, X.MODEL, K = 40, pred.fun)
}
\arguments{
\item{X}{the data used to train the model}

\item{K}{the number of bins split to evaluate the ALE Plot see ALEPLOT package}

\item{pred.fun}{the prediction function required by ALEPLOT package see details}

\item{X, MODEL}{the model}
}
\value{
a list of ALEplots ordered by variable importance.
}
\description{
ALEPlots are a method to visualize the relationship between individual predictors and an outcome within ordinarily
difficult to interpret non-linear machine learning models. This function produces a list of plots with an aleplot
for each predictor within an ML model.
}
