% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEplotPlus.R
\name{pltROC}
\alias{pltROC}
\title{pltROC}
\usage{
pltROC(predictions, truevals)
}
\arguments{
\item{predictions}{predictions for the ROC as a probability 0 to 1}

\item{truevals}{binary factor make sure levels are aligned properly}
}
\value{
a list with the plot and the auc info
}
\description{
Plots an ROC with an AUC with all the points labeled and an x / y abline
}
