% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareTestData.R
\name{SetupMtcars}
\alias{SetupMtcars}
\title{SetupMTCars}
\usage{
SetupMtcars()
}
\value{
A list containing the following elements:
\describe{
  \item{X}{A data frame of the mtcars data without the \code{mpg} column.}
  \item{X.MODEL}{A random forest model object fitted to the mtcars data.}
  \item{pred.fun}{A function to make predictions using the random forest model.}
}
}
\description{
This function sets up the mtcars dataset, which contains various predictors for miles per gallon (mpg).
}
\details{
The dataset includes the following predictors:
\itemize{
  \item \strong{mpg} - Miles/(US) gallon.
  \item \strong{cyl} - Number of cylinders.
  \item \strong{disp} - Displacement (cu.in.).
  \item \strong{hp} - Gross horsepower.
  \item \strong{drat} - Rear axle ratio.
  \item \strong{wt} - Weight (1000 lbs).
  \item \strong{qsec} - 1/4 mile time.
  \item \strong{vs} - Engine (0 = V-shaped, 1 = straight).
  \item \strong{am} - Transmission (0 = automatic, 1 = manual).
  \item \strong{gear} - Number of forward gears.
  \item \strong{carb} - Number of carburetors.
}

The target variable is:
\itemize{
  \item \strong{mpg} - Miles/(US) gallon.
}
}
