% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareTestData.R
\name{SetupPimi}
\alias{SetupPimi}
\title{SetupPima}
\usage{
SetupPimi()
}
\value{
A list containing the following elements:
\describe{
  \item{X}{A data frame of the Pima Indian Diabetes data without the \code{Outcome} column.}
  \item{X.MODEL}{A random forest model object fitted to the Pima Indian Diabetes data.}
  \item{pred.fun}{A function to make predictions using the random forest model.}
}
}
\description{
This function sets up the Pima Indian Diabetes Dataset, which contains various predictors for diabetes.
}
\details{
The dataset includes the following predictors:
\itemize{
  \item \strong{Pregnancies} - Number of times pregnant.
  \item \strong{Glucose} - Plasma glucose concentration after 2 hours in an oral glucose tolerance test.
  \item \strong{BloodPressure} - Diastolic blood pressure (mm Hg).
  \item \strong{SkinThickness} - Triceps skin fold thickness (mm).
  \item \strong{Insulin} - 2-hour serum insulin (mu U/ml).
  \item \strong{BMI} - Body mass index (weight in kg/(height in m)^2).
  \item \strong{DiabetesPedigreeFunction} - Diabetes pedigree function (a function which scores likelihood of diabetes based on family history).
  \item \strong{Age} - Age in years.
}

The target variable is:
\itemize{
  \item \strong{Outcome} - Binary variable indicating if the patient has diabetes (1) or not (0).
}
}
