% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEplotPlus.R
\name{str_safe}
\alias{str_safe}
\title{str_safe}
\usage{
str_safe(strng)
}
\arguments{
\item{strng}{string to make safe}
}
\value{
a string with the dangerous parts replaced but still looks very similar to a human reader
}
\description{
This function replaces special symbols with close approximations that dont mess up file systems
}
